requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
cffi==2.0.0
click==8.3.1
cloudpickle==3.1.2
Django==5.2.8
django-jazzmin==3.0.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.110.0
h11==0.16.0
idna==3.11
imbalanced-learn==0.11.0
joblib==1.3.2
llvmlite==0.45.1
numba==0.62.1
numpy==1.26.4
packaging==25.0
pandas==2.0.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.15.3
shap==0.44.1
six==1.17.0
slicer==0.0.7
sniffio==1.3.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.36.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.27.0
xgboost==1.7.6
